Move queryParam to hook package
