+++ frontend/packages/hooks/package.json
@@ -25,7 +25,8 @@
     },
     "peerDependencies": {
         "react": "^18.3.1",
-        "react-dom": "^18.3.1"
+        "react-dom": "^18.3.1",
+        "react-router-dom": "^6.25.1"
     },
     "devDependencies": {
         "@types/react": "^18.3.18",
