+++ frontend/packages/hooks/src/index.ts
@@ -5,3 +5,4 @@ export { useGetFrappeDoctypeMeta } from "./useGetFrappeDoctypeMeta";
 export { useGetFrappeDocTypeCount } from "./useGetFrappeDocTypeCount";
 export { usePagination } from "./usePagination";
 export { useInfiniteScroll } from "./useInfiniteScroll";
+export { useQueryParam } from "./useQueryParam";
